Document online consensus
At the last CG meeting I was asked to make sure that we could make forward progress online. We also intend to hold semi-frequent video chats for the CG. Folks were concerned that the consensus process we documented for in-person meetings was too onerous in-between meetings, I'm therefore proposing that we adopt this online process. Talking to our standards folks this is usually left undocumented, but they like that this is precise and therefore unambiguous.