Add code of conduct mention
This PR adds a link to Igalia's standard events code of conduct.  It
happens to link out to a page from a later event but to avoid importing
the specific CoC into this repo, and to avoid the question of what the
W3C's / WebAssembly CG's official CoC is, this seems like a lightweight
solution.

This PR also adds a time in the morning to briefly remind people of the
CoC and of host organization contacts for reporting any problem.