Discuss: It's hard to contribute on proposals.
I find it difficult to contribute to existing proposals. They are a forked WASI repos with additional files.

There's no space for discussion on them, other than creating a PR that targets the fork. I'd like to have a space to discuss first (Github discussions, maybe?!?). Even, opening an PR is hard as it may interfere with the main maintainer commits.

I think the better might be to commit (by PR) a stub of a proposal first. So, then multiple people could provide parts of the proposal in smaller chunks.