Document that phase advancement is determined by CG vote. (#1312)
I noticed that CG voting was only mentioned as an entry requirement to
phase 1. In practice, the CG has always voted for all phase transitions
beyond phase 0.

This PR documents this existing practice.

It adds a paragraph documenting that votes are used to advance from one
phase to another, and it moves the phase 1 entry requirements out of
phase 0's description and into a phase 1 entry requirements section,
making it more consistent with the other phases.