Add poll for wasi-backtrace
In the last meeting, I heard some support for this idea so let's put it
up for a vote. A couple of people (Luke, Syrus) suggested that
wasi-backtrace could or should be a core level proposal; I agree that
the core spec _could_ be a good place, but WASI's recent change to
become a collection of "standard interfaces" instead of "system
interfaces" suggests WASI as a suitable place as well. Additionally,
moving this to phase 1 makes it possible to prototype with this API
under a flag and sort out the details before we need to decide exactly
where this needs to live. As support improves for languages that reflect
stack traces in errors, we should soon hear whether what was presented
here is a good fit.

Note that @tlively floated a similar idea to this almost exactly four
years ago in [WASI#159].

[WASI#159]: https://github.com/WebAssembly/WASI/issues/159