Add custom Wasm page sizes agenda item to CG-03-26.md (#1511)
I will make a quick presentation giving an overview of the proposed design and some of the discussion that's happened on the issue thus far. Seeking broader CG feedback. Unsure about 30 minutes, I expect the presentation will be faster than that but I want to leave room for actual discussion.