Copy code of conduct to meetings repo
Copy the WebAssembly code of conduct, currently hosted at
https://github.com/WebAssembly/design/blob/main/CodeOfConduct.md, from the
design repo to the meetings repo.

The documents in the design repo are largely out of date and of historical
interest only, while the documents in the meetings repo are current, so the
meetings repo is a much better place for the code of conduct.

Add the code of conduct in a top-level CODE_OF_CONDUCT file instead of in the
process directory to 1) make it even more visible, and 2) take advantage of
GitHub's native support for codes of conduct.

If this seems like a good direction, the next step would be to update other
repos in our organization with code of conduct files that redirect to this
canonical file.