Update consensus.md to reflect formal and informal consensus
Consensus.md previously drew a distinction between in-person and online
consensus, but this did not accurately represent the status quo where
online CG meetings are just as official as in-person CG meetings. Update
the document to instead draw a distinction between "informal" consensus
(i.e. consensus that does not require a poll) and formal consensus.

Move the section on informal consensus to be first, since it describes
when formal consensus is necessary or not.

This change is similar to and subsumes #1418.