Add a presentation on Lime1. (#1839)
[Lime] is the name of a new series of stable feature sets defined in the
[tool conventions] repository, currently with one set: [Lime1]. The goal
is to give engines and users a stable target that has more features than
the MVP, which is what often gets used instead. I recently heard that
not everyone has heard about it yet, so I'd like to give a very brief
resentation to the CG.

I'm proposing to insert this presentation before the JS interop topic,
because I'm expecting it'll be quick. If people want to ask more
questions or dive into how this might to profiles or so, I'll propose we
take it offline or schedule time in a later meeting.

[Lime]:
https://github.com/WebAssembly/tool-conventions/blob/main/Lime.md
[tool conventions]: https://github.com/WebAssembly/tool-conventions
[Lime1]:
https://github.com/WebAssembly/tool-conventions/blob/main/Lime.md#lime1